requirements.txt
flask
tensorflow
numpy
Pillow
gradio_client